#105 - Refactor of RandomTool to package
