+++ src/main/java/net/coru/kloadgen/extractor/extractors/AvroExtractor.java
@@ -10,7 +10,7 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Set;
 import net.coru.kloadgen.model.FieldValueMapping;
-import net.coru.kloadgen.util.RandomTool;
+import net.coru.kloadgen.randomtool.random.RandomObject;
 import org.apache.avro.Schema;
 
 public class AvroExtractor {
@@ -21,6 +21,12 @@ public class AvroExtractor {
 
   private final Set<Schema.Type> typesSet = EnumSet.of(Schema.Type.INT, Schema.Type.DOUBLE, Schema.Type.FLOAT, Schema.Type.BOOLEAN, Schema.Type.STRING,
                                                        Schema.Type.LONG, Schema.Type.BYTES, Schema.Type.FIXED);
+
+  private final RandomObject randomObject;
+
+  public AvroExtractor() {
+    randomObject = new RandomObject();
+  }
 
   public List<FieldValueMapping> processSchema(Schema schema) {
     List<FieldValueMapping> attributeList = new ArrayList<>();
@@ -141,7 +147,7 @@ public class AvroExtractor {
     chosenType = extractTypeName(types.get(1)).equalsIgnoreCase("array") ? types.get(1) : chosenType;
     String chosenTypeName = extractTypeName(chosenType);
 
-    if (!RandomTool.isTypeValid(chosenTypeName)) {
+    if (!randomObject.isTypeValid(chosenTypeName)) {
       chosenTypeName = "null";
     } else if ("array".equalsIgnoreCase(chosenTypeName)) {
       chosenTypeName = "int-array";
