+++ src/test/java/net/coru/kloadgen/randomtool/generator/AvroGeneratorToolTest.java
@@ -1,4 +1,4 @@
-package net.coru.kloadgen.util;
+package net.coru.kloadgen.randomtool.generator;
 
 import static java.util.Collections.emptyList;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -17,13 +17,14 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class AvroRandomToolTest {
+class AvroGeneratorToolTest {
 
   private static final LocalDateTime FIXED_DATE = LocalDateTime.of(2019, 12, 6, 12, 0, 0);
 
   private static Stream<Arguments> parametersForGenerateRandomValueForField() {
     return Stream.of(
-        Arguments.of("string", 1, Collections.singletonList("testString"), new Field("name", SchemaBuilder.builder().stringType()), "testString"),
+        Arguments.of("string", 1, Collections.singletonList("testString"), new Field("name", SchemaBuilder.builder().stringType()),
+            "testString"),
         Arguments.of("string", 1, Collections.singletonList("1"), new Field("name", SchemaBuilder.builder().intType()), 1),
         Arguments.of("int", 1, Collections.singletonList("1"), new Field("name", SchemaBuilder.builder().stringType()), "1"),
         Arguments.of("int", 1, Collections.singletonList("1"), new Field("name", SchemaBuilder.builder().intType()), 1),
@@ -31,24 +32,22 @@ class AvroRandomToolTest {
         Arguments.of("short", 1, Collections.singletonList("1"), new Field("name", SchemaBuilder.builder().intType()), (short) 1),
         Arguments.of("double", 1, Collections.singletonList("1.0"), new Field("name", SchemaBuilder.builder().doubleType()), 1.0),
         Arguments.of("float", 1, Collections.singletonList("1.0"), new Field("name", SchemaBuilder.builder().floatType()), 1.0f),
-        Arguments.of("timestamp", 1, Collections.singletonList("2019-12-06T12:00:00"), new Field("name", SchemaBuilder.builder().stringType()),
-            FIXED_DATE),
-        Arguments.of("longTimestamp", 1, Collections.singletonList("2019-12-06T12:00:00"), new Field("name", SchemaBuilder.builder().longType()),
-            FIXED_DATE.toInstant(ZoneOffset.UTC).toEpochMilli()),
-        Arguments.of("stringTimestamp", 1, Collections.singletonList("2019-12-06T12:00:00"), new Field("name", SchemaBuilder.builder().stringType()),
+        Arguments
+            .of("timestamp", 1, Collections.singletonList("2019-12-06T12:00:00"), new Field("name", SchemaBuilder.builder().stringType()),
+                FIXED_DATE),
+        Arguments
+            .of("longTimestamp", 1, Collections.singletonList("2019-12-06T12:00:00"), new Field("name", SchemaBuilder.builder().longType()),
+                FIXED_DATE.toInstant(ZoneOffset.UTC).toEpochMilli()),
+        Arguments.of("stringTimestamp", 1, Collections.singletonList("2019-12-06T12:00:00"),
+            new Field("name", SchemaBuilder.builder().stringType()),
             "2019-12-06T12:00"),
         Arguments.of("uuid", 1, Collections.singletonList("0177f035-e51c-4a46-8b82-5b157371c2a5"),
             new Field("name", SchemaBuilder.builder().stringType()), UUID.fromString("0177f035-e51c-4a46-8b82-5b157371c2a5")),
-        Arguments.of("boolean", 1, Collections.singletonList("true"), new Field("name", SchemaBuilder.builder().booleanType()), Boolean.TRUE),
+        Arguments
+            .of("boolean", 1, Collections.singletonList("true"), new Field("name", SchemaBuilder.builder().booleanType()), Boolean.TRUE),
         Arguments.of("boolean", 1, Collections.singletonList("true"), new Field("name", SchemaBuilder.builder().stringType()), "true"),
         Arguments
             .of("string", 1, Collections.singletonList("true"), new Field("name", SchemaBuilder.builder().booleanType()), Boolean.TRUE));
-  }
-
-  @ParameterizedTest
-  @MethodSource("parametersForGenerateRandomValueForField")
-  void testGenerateRandomValueForField(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field, Object expected) {
-    assertThat(new AvroRandomTool().generateRandom(fieldType, valueLength, fieldValuesList, field)).isEqualTo(expected);
   }
 
   private static Stream<Arguments> parametersForGenerateRandomValue() {
@@ -58,25 +57,10 @@ class AvroRandomToolTest {
         Arguments.of("double", 11, emptyList(), new Field("name", SchemaBuilder.builder().doubleType())));
   }
 
-  @ParameterizedTest
-  @MethodSource("parametersForGenerateRandomValue")
-  void testGenerateRandomValue(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field) {
-    Object number = new AvroRandomTool().generateRandom(fieldType, valueLength, fieldValuesList, field);
-    assertThat(number).isInstanceOfAny(Long.class, Integer.class, Double.class, Float.class);
-    assertThat(String.valueOf(number)).hasSize(valueLength);
-  }
-
   private static Stream<Arguments> parametersForGenerateRandomValueForEnums() {
     return Stream.of(
         Arguments.of("enum", 1, Collections.singletonList("RED"),
             new Field("name", SchemaBuilder.builder().enumeration("ENUM1").symbols("RED", "BLUE", "GREEN")), "RED"));
-  }
-
-  @ParameterizedTest
-  @MethodSource("parametersForGenerateRandomValueForEnums")
-  void testGenerateRandomValueForEnums(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field, Object expected) {
-    assertThat(new AvroRandomTool().generateRandom(fieldType, valueLength, fieldValuesList, field))
-        .hasFieldOrPropertyWithValue("symbol", expected);
   }
 
   private static Stream<Arguments> parametersForGenerateSequenceValueForField() {
@@ -84,13 +68,6 @@ class AvroRandomToolTest {
         Arguments.of("seq", 1, Collections.singletonList("0"), new Field("name", SchemaBuilder.builder().stringType()), "0"),
         Arguments.of("seq", 1, Collections.singletonList("1"), new Field("name", SchemaBuilder.builder().intType()), 1),
         Arguments.of("seq", 1, Collections.singletonList("2"), new Field("name", SchemaBuilder.builder().intType()), 2));
-  }
-
-  @ParameterizedTest
-  @MethodSource("parametersForGenerateSequenceValueForField")
-  void testGenerateSequenceValueForField(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field,
-      Object expectedTyped) {
-    assertThat(new AvroRandomTool().generateRandom(fieldType, valueLength, fieldValuesList, field)).isEqualTo(expectedTyped);
   }
 
   private static Stream<Arguments> parametersForShouldRecoverVariableFromContext() {
@@ -112,12 +89,40 @@ class AvroRandomToolTest {
   }
 
   @ParameterizedTest
+  @MethodSource("parametersForGenerateRandomValueForField")
+  void testGenerateRandomValueForField(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field, Object expected) {
+    assertThat(new AvroGeneratorTool().generateObject(field, fieldType, valueLength, fieldValuesList)).isEqualTo(expected);
+  }
+
+  @ParameterizedTest
+  @MethodSource("parametersForGenerateRandomValue")
+  void testGenerateRandomValue(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field) {
+    Object number = new AvroGeneratorTool().generateObject(field, fieldType, valueLength, fieldValuesList);
+    assertThat(number).isInstanceOfAny(Long.class, Integer.class, Double.class, Float.class);
+    assertThat(String.valueOf(number)).hasSize(valueLength);
+  }
+
+  @ParameterizedTest
+  @MethodSource("parametersForGenerateRandomValueForEnums")
+  void testGenerateRandomValueForEnums(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field, Object expected) {
+    assertThat(new AvroGeneratorTool().generateObject(field, fieldType, valueLength, fieldValuesList))
+        .hasFieldOrPropertyWithValue("symbol", expected);
+  }
+
+  @ParameterizedTest
+  @MethodSource("parametersForGenerateSequenceValueForField")
+  void testGenerateSequenceValueForField(String fieldType, Integer valueLength, List<String> fieldValuesList, Field field,
+      Object expectedTyped) {
+    assertThat(new AvroGeneratorTool().generateObject(field, fieldType, valueLength, fieldValuesList)).isEqualTo(expectedTyped);
+  }
+
+  @ParameterizedTest
   @MethodSource("parametersForShouldRecoverVariableFromContext")
   void shouldRecoverVariableFromContext(String fieldType, Integer valueLength, String value, Field field, Object expected) {
     JMeterVariables variables = new JMeterVariables();
     variables.put("VARIABLE", value);
     JMeterContextService.getContext().setVariables(variables);
-    assertThat(new AvroRandomTool().generateRandom(fieldType, valueLength, Collections.singletonList("${VARIABLE}"), field))
+    assertThat(new AvroGeneratorTool().generateObject(field, fieldType, valueLength, Collections.singletonList("${VARIABLE}")))
         .isEqualTo(expected);
   }
 }
