+++ src/test/java/net/coru/kloadgen/randomtool/random/RandomArrayTest.java
@@ -1,0 +1,34 @@
+package net.coru.kloadgen.randomtool.random;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Stream;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+class RandomArrayTest {
+
+  private static Stream<Arguments> parametersForGenerateArrayRandomValue() {
+    return Stream.of(
+        Arguments.of("string-array", 1, Collections.singletonList("testString"), "testString"),
+        Arguments.of("int-array", 1, Collections.singletonList("1"), 1),
+        Arguments.of("long-array", 1, Collections.singletonList("1"), 1L),
+        Arguments.of("short-array", 1, Collections.singletonList("1"), (short) 1),
+        Arguments.of("double-array", 1, Collections.singletonList("1.0"), 1.0),
+        Arguments.of("float-array", 1, Collections.singletonList("1.0"), 1.0f),
+        Arguments.of("uuid-array", 1, Collections.singletonList("0177f035-e51c-4a46-8b82-5b157371c2a5"),
+            UUID.fromString("0177f035-e51c-4a46-8b82-5b157371c2a5"))
+    );
+  }
+
+  @ParameterizedTest
+  @MethodSource("parametersForGenerateArrayRandomValue")
+  void generateArrayRandomValue(String fieldType, Integer valueLength, List<String> fieldValuesList, Object expected) {
+    assertThat((List<Object>) new RandomArray().generateArray(fieldType, valueLength, fieldValuesList, 1, Collections.emptyMap()))
+        .allMatch(value -> value.equals(expected));
+  }
+}
